pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Bosta simples pro meu irmão"
authors = [
    {name = "Gabriel",email = "gabrielmsol37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "shareplum (>=0.5.1,<0.6.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.63.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"